[REFACTOR] [SWE-663] Refactored customer page, created Slider component #done
