+++ components/elements/Slider/index.js
@@ -1,0 +1,3 @@
+import Slider from './Slider'
+
+export default Slider
