+++ components/elements/index.js
@@ -1,5 +1,5 @@
 export { default as Button } from './Button'
-export { default as CustomerSlider } from './CustomerSlider'
+export { default as Slider } from './Slider'
 export { default as SectionNavigationDots } from './SectionNavigationDots'
 export { default as Stepper } from './Stepper'
 export { default as Text } from './Typography/Text'
