+++ domains/Customer/__constants__/index.js
@@ -1,0 +1,1 @@
+export { default as CUSTOMERS_SLIDES } from './customersSlides'
