+++ domains/Customer/components/CustomerSlide/index.js
@@ -1,0 +1,3 @@
+import CustomerSlide from './CustomerSlide'
+
+export default CustomerSlide
