+++ domains/Customer/components/CustomerSlider/index.js
