+++ domains/Customer/components/index.js
@@ -1,0 +1,2 @@
+export { default as CustomerSlider } from './CustomerSlider'
+export { default as CustomerSlide } from './CustomerSlide'
