+++ public/locales/en/landing.json
@@ -343,7 +343,7 @@
         "subtitle": "Best price/quality ratio"
       }
     },
-    "customers_slides_items": {
+    "customers_slides": {
       "menta": {
         "description": "Menta Health is an international initiative to improve and streamline the treatment of anxiety and depression.",
         "aria": "Go to Menta slide"
