+++ public/locales/no/landing.json
@@ -374,7 +374,7 @@
         "subtitle": "Beste forhold mellom pris og kvalitet"
       }
     },
-    "customers_slides_items": {
+    "customers_slides": {
       "menta": {
         "description": "Menta Helse er en internasjonal satsing for å forbedre og effektivisere behandling av angst - og depresjonslidelser.",
         "aria": "Gå til Menta slide"
