--- styles/components/customersSlider.css
@@ -1,140 +1,0 @@
-.customers-slider-wrapper {
-  position: relative;
-}
-.customers-slider-wrapper .cover-gradient {
-  position: absolute;
-  bottom: 0;
-  left: 0;
-  right: 0;
-  background-image: -webkit-linear-gradient(
-    top,
-    rgba(0, 0, 0, 0),
-    rgba(0, 0, 0, 1)
-  );
-  height: 50vh;
-}
-
-.customers-slider-wrapper .arrows-wrapper {
-  position: absolute;
-  top: calc(100% / 2 - 12px);
-  display: flex;
-  width: 100%;
-  justify-content: space-between;
-  padding: 0 10px;
-}
-
-.customers-slider-wrapper .arrows-wrapper .prev-slide-icon-wrapper,
-.customers-slider-wrapper .arrows-wrapper .next-slide-icon-wrapper {
-  width: 24px;
-  height: 24px;
-  background-color: rgba(0, 0, 0, 0.3);
-  border-radius: 50%;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  cursor: pointer;
-  transition: all 0.6s ease;
-  position: relative;
-  z-index: 1;
-}
-.customers-slider-wrapper .prev-slide-icon-wrapper:hover,
-.customers-slider-wrapper .next-slide-icon-wrapper:hover {
-  background-color: rgba(0, 0, 0, 0.6);
-  transition: all 0.6s ease;
-}
-
-.customers-slider-wrapper .prev-slide-icon,
-.customers-slider-wrapper .next-slide-icon {
-  width: 4px;
-  height: 8px;
-  position: relative;
-}
-.customers-slider-wrapper .slider-item {
-  height: 85vh;
-}
-
-.customers-slider-wrapper .content-wrapper {
-  color: var(--white);
-  padding: 0 40px 70px 40px;
-  height: 100%;
-  display: flex;
-  flex-direction: column;
-  justify-content: flex-end;
-  align-items: center;
-  position: relative;
-}
-.customers-slider-wrapper .content-wrapper .content {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-}
-.customers-slider-wrapper .description {
-  text-align: center;
-  margin-bottom: 48px;
-}
-.customers-slider-wrapper .go-to-site {
-  width: fit-content;
-}
-.customers-slider-wrapper .dots-wrapper {
-  position: absolute;
-  bottom: 0;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  width: 100%;
-}
-
-.customers-slider-wrapper .dots-wrapper .dots {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  padding-bottom: 26px;
-  z-index: 10;
-}
-.customers-slider-wrapper .dots-wrapper .dots .dot {
-  cursor: pointer;
-  height: 10px;
-  width: 10px;
-  margin: 0 6px;
-  border: 2px solid var(--white);
-  border-radius: 50%;
-  display: inline-block;
-  transition: background-color 1s ease;
-}
-
-.customers-slider-wrapper .dot.active,
-.customers-slider-wrapper .dot.active:hover {
-  background-color: var(--white);
-}
-
-@media (min-width: 576px) {
-  .customers-slider-wrapper .arrows-wrapper {
-    padding: 0 20px;
-  }
-  .customers-slider-wrapper .arrows-wrapper .prev-slide-icon-wrapper,
-  .customers-slider-wrapper .arrows-wrapper .next-slide-icon-wrapper {
-    width: 40px;
-    height: 40px;
-    top: calc(100% / 2 - 20px);
-  }
-  .customers-slider-wrapper .prev-slide-icon,
-  .customers-slider-wrapper .next-slide-icon {
-    width: 12px;
-    height: 18px;
-  }
-
-  .customers-slider-wrapper .content-wrapper .content {
-    max-width: 560px;
-  }
-}
-@media (min-width: 960px) {
-  .customers-slider-wrapper .content-wrapper .content {
-    max-width: 460px;
-  }
-}
-
-@media (min-width: 1200px) {
-  .customers-slider-wrapper .content-wrapper .content {
-    max-width: 560px;
-  }
-}
