+++ styles/components/index.css
@@ -6,7 +6,7 @@
 @import url("./button.css");
 @import url("./menu.css");
 @import url("./languagesMenu.css");
-@import url("./customersSlider.css");
+@import url("./slider.css");
 @import url("./sectionNavigationDots.css");
 @import url("./stepper.css");
 @import url("./projectCount.css");
