+++ styles/components/slider.css
@@ -1,0 +1,100 @@
+.slider-wrapper {
+  position: relative;
+}
+.slider-wrapper .cover-gradient {
+  position: absolute;
+  bottom: 0;
+  left: 0;
+  right: 0;
+  background-image: -webkit-linear-gradient(
+    top,
+    rgba(0, 0, 0, 0),
+    rgba(0, 0, 0, 1)
+  );
+  height: 50vh;
+}
+
+.slider-wrapper .arrows-wrapper {
+  position: absolute;
+  top: calc(100% / 2 - 12px);
+  display: flex;
+  width: 100%;
+  justify-content: space-between;
+  padding: 0 10px;
+}
+
+.slider-wrapper .arrows-wrapper .prev-slide-icon-wrapper,
+.slider-wrapper .arrows-wrapper .next-slide-icon-wrapper {
+  width: 24px;
+  height: 24px;
+  background-color: rgba(0, 0, 0, 0.3);
+  border-radius: 50%;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  cursor: pointer;
+  transition: all 0.6s ease;
+  position: relative;
+  z-index: 1;
+}
+.slider-wrapper .prev-slide-icon-wrapper:hover,
+.slider-wrapper .next-slide-icon-wrapper:hover {
+  background-color: rgba(0, 0, 0, 0.6);
+  transition: all 0.6s ease;
+}
+
+.slider-wrapper .prev-slide-icon,
+.slider-wrapper .next-slide-icon {
+  width: 4px;
+  height: 8px;
+  position: relative;
+}
+
+.slider-wrapper .dots-wrapper {
+  position: absolute;
+  bottom: 0;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  width: 100%;
+}
+
+.slider-wrapper .dots-wrapper .dots {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  padding-bottom: 26px;
+  z-index: 10;
+}
+.slider-wrapper .dots-wrapper .dots .dot {
+  cursor: pointer;
+  height: 10px;
+  width: 10px;
+  margin: 0 6px;
+  border: 2px solid var(--white);
+  border-radius: 50%;
+  display: inline-block;
+  transition: background-color 1s ease;
+}
+
+.slider-wrapper .dot.active,
+.slider-wrapper .dot.active:hover {
+  background-color: var(--white);
+}
+
+@media (min-width: 576px) {
+  .slider-wrapper .arrows-wrapper {
+    padding: 0 20px;
+  }
+  .slider-wrapper .arrows-wrapper .prev-slide-icon-wrapper,
+  .slider-wrapper .arrows-wrapper .next-slide-icon-wrapper {
+    width: 40px;
+    height: 40px;
+    top: calc(100% / 2 - 20px);
+  }
+  .slider-wrapper .prev-slide-icon,
+  .slider-wrapper .next-slide-icon {
+    width: 12px;
+    height: 18px;
+  }
+}
