+++ styles/domains/customer.css
@@ -1,0 +1,42 @@
+.slider-wrapper .customer-slide-wrapper {
+  height: 85vh;
+}
+
+.slider-wrapper .customer-slide-wrapper .content-wrapper {
+  color: var(--white);
+  padding: 0 40px 70px 40px;
+  height: 100%;
+  display: flex;
+  flex-direction: column;
+  justify-content: flex-end;
+  align-items: center;
+  position: relative;
+}
+.slider-wrapper .customer-slide-wrapper .content-wrapper .content {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+}
+.slider-wrapper .customer-slide-wrapper .description {
+  text-align: center;
+  margin-bottom: 48px;
+}
+.slider-wrapper .customer-slide-wrapper .go-to-site {
+  width: fit-content;
+}
+@media (min-width: 576px) {
+  .slider-wrapper .customer-slide-wrapper .content-wrapper .content {
+    max-width: 560px;
+  }
+}
+@media (min-width: 960px) {
+  .slider-wrapper .customer-slide-wrapper .content-wrapper .content {
+    max-width: 460px;
+  }
+}
+
+@media (min-width: 1200px) {
+  .slider-wrapper .customer-slide-wrapper .content-wrapper .content {
+    max-width: 560px;
+  }
+}
