+++ styles/domains/index.css
@@ -10,3 +10,4 @@
 @import url("./technology.css");
 @import url("./package.css");
 @import url("./contact.css");
+@import url("./customer.css");
